+++ doc/changelog.rst
@@ -56,6 +56,11 @@ Highlights include:
   default. Users that require this behavior are encouraged to upgrade to
   CPython 3.4+.
 
+Since 3.4rc0, the max staleness option has been renamed from ``maxStalenessMS``
+to ``maxStalenessSeconds``, its smallest value has changed from twice
+``heartbeatFrequencyMS`` to ``heartbeatFrequencyMS`` plus 10 seconds, and its
+default value has changed from ``None`` or 0 to -1.
+
 .. _PEP 446: https://www.python.org/dev/peps/pep-0446/
 
 Issues Resolved